Restructured types
