+++ src/types.ts
@@ -5,80 +5,6 @@
 export const ALXmlDocDiagnosticPrefix = "AL-XML-DOC";
 
 /**
- * AL Object Types.
- */
-export enum ALObjectType {
-	Unknown,
-	Codeunit,
-	Table,
-    Page,
-	Query,
-	Report,
-	XmlPort,
-	Enum,
-	TableExtension,
-	PageExtension,
-	EnumExtension,
-	Interface,
-	ControlAddIn
-};
-
-/**
- * AL Object Extension Type
- */
-export enum ALObjectExtensionType {
-	Extend,
-	Implement
-}
-
-/**
- * AL Codeunit Types.
- */
-export enum ALCodeunitType {
-	Normal,
-	Test,
-	TestRunner,
-	Install,
-	Upgrade
-}
-
-/**
- * AL Access Level.
- */
-export enum ALAccessLevel {
-	Public,
-	Internal,
-	Local,
-	Protected
-}
-
-/**
- * AL Obsolete State.
- */
-export enum ALObsoleteState {
-	No,
-	Pending
-}
-
-/**
- * AL Procedure Types.
- */
-export enum ALProcedureType {
-	Procedure,
-	Trigger
-}
-
-/**
- * AL Procedure Subtype.
- */
-export enum ALProcedureSubtype {
-	Normal,
-	Test,
-	EventPublisher,
-	EventSubscriber
-}
-
-/**
  * Diagnostics codes.
  */
 export enum ALXmlDocDiagnosticCode {
