+++ src/types/ALAccessLevel.ts
@@ -1,0 +1,9 @@
+/**
+ * AL Access Level.
+ */
+export enum ALAccessLevel {
+	Public,
+	Internal,
+	Local,
+	Protected
+}
