+++ src/types/ALCodeunitType.ts
@@ -1,0 +1,10 @@
+/**
+ * AL Codeunit Types.
+ */
+export enum ALCodeunitType {
+	Normal,
+	Test,
+	TestRunner,
+	Install,
+	Upgrade
+}
