+++ src/types/ALObjectExtensionType.ts
@@ -1,0 +1,7 @@
+/**
+ * AL Object Extension Type
+ */
+export enum ALObjectExtensionType {
+	Extend,
+	Implement
+}
