+++ src/types/ALObjectType.ts
@@ -1,0 +1,18 @@
+/**
+ * AL Object Types.
+ */
+export enum ALObjectType {
+	Unknown,
+	Codeunit,
+	Table,
+    Page,
+	Query,
+	Report,
+	XmlPort,
+	Enum,
+	TableExtension,
+	PageExtension,
+	EnumExtension,
+	Interface,
+	ControlAddIn
+};
