+++ src/types/ALObsoleteState.ts
@@ -1,0 +1,8 @@
+/**
+ * AL Obsolete State.
+ */
+export enum ALObsoleteState {
+	No,
+    Pending,
+    Removed
+}
