+++ src/types/ALProcedureSubtype.ts
@@ -1,0 +1,9 @@
+/**
+ * AL Procedure Subtype.
+ */
+export enum ALProcedureSubtype {
+	Normal,
+	Test,
+	EventPublisher,
+	EventSubscriber
+}
