+++ src/types/ALProcedureType.ts
@@ -1,0 +1,7 @@
+/**
+ * AL Procedure Types.
+ */
+export enum ALProcedureType {
+	Procedure,
+	Trigger
+}
